training/requirements.txt
google-cloud-storage==2.18.2
google-cloud-aiplatform==1.65.0
langchain==0.2.16
langchain-community==0.2.16
langchain-google-vertexai==1.0.10
qdrant-client==1.11.1
pypdf==3.17.4
click==8.1.7
Flask==2.3.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
Werkzeug==3.0.6
kubernetes==28.1.0
pika==1.3.2
